Merge equalisation and contrasting in one command
Speeds things up!
The result is very similar to the previous implementation but not exactly equal. Indeed the lightening factor is applied to the original maximum while in the previous case it was applied to the maximum after subtracting the min. Perceptually, it does not make much of a difference.