Re-write image lightening and thresholding
Grey rescaling based on min and max is sensitive to extreme values in some images. It needs to be done based on percentiles.

Percentiles were already computed for the segmentation. Move everything outside of the segment() function and do it only once.

As a side effect, to inspect the influence of those parameters, it is now possible to save images before and after processing. All images are saved in the same directory, with suffixes.